requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.5
django-cors-headers==3.7.0
django-tenants==3.3.2
djangorestframework==3.12.4
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
requests==2.25.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.6
zope.interface==5.4.0
